evas/cserve2_slave: Fix build race with linking to eo but not depending on it.
I have not been able to reproduce this myself but I have seen a build log where
the binary tries to link to libeo and fails due to the missing file.
A similar problem was "fixed" in 0e4b847deb79c957dbc042ea0c6a98282aead6e0, but
this really makes me wonder where the linking against eo comes from for cserve2
which is not using eo as far as I can see.