evg: add Vector library abstraction.
EVG is mean to be a higher level abstraction that will enable the writing of
widget using a description of it in a vector form. This is necessary to extend
SVG to be usable to design widget. There will always be a cost to do widget
using vector graphics, but for some design, it is the only way to do it. EFL
has always wanted to provide every possible design to its user, so it is
natural to add it here.

EVG will use Evas_Object_VG (the SVG scenegraph widget) to provide this infrastructure.
It will manipulate the Evas_VG_Node to scale correctly in the context of a widget
which is not the same as an image like SVG. The idea for doing the design a vector
widget will be to extend an SVG file by providing relative positionning information.